# Optimize log4net log injection and refactor log providers