--- a/src/Datadog.Trace/Logging/ILogProviderWithEnricher.cs
+++ b/src/Datadog.Trace/Logging/ILogProviderWithEnricher.cs
@@ -1,0 +1,12 @@
+// <copyright file="ILogProviderWithEnricher.cs" company="Datadog">
+// Unless explicitly stated otherwise all files in this repository are licensed under the Apache 2 License.
+// This product includes software developed at Datadog (https://www.datadoghq.com/). Copyright 2017 Datadog, Inc.
+// </copyright>
+
+namespace Datadog.Trace.Logging
+{
+    internal interface ILogProviderWithEnricher
+    {
+        ILogEnricher CreateEnricher();
+    }
+}