log4net benchmark:

|      Method |     Mean |   Error |  StdDev |  Gen 0 | Gen 1 | Gen 2 | Allocated |
|------------ |---------:|--------:|--------:|-------:|------:|------:|----------:|
|         Old | 127.0 us | 2.07 us | 1.93 us | 0.2441 |     - |     - |  27.71 KB |
|         New | 116.2 us | 1.96 us | 1.64 us |      - |     - |     - |   9.91 KB |